Calculate clock correlation distance (CCD).

calcCCD quantifies the similarity of gene co-expression between a reference and a test dataset. Statistical significance is calculated using permutation of the genes.

calcCCD(refCor, emat, groupVec = NULL, refEmat = NULL, nPerm = 1000,
  geneNames = NULL, dopar = FALSE)

Arguments

refCor
Correlation matrix to be used as the reference, such as comes from getRefCor(). Should contain Spearman correlation values.

emat
Matrix of expression values, where each row corresponds to a gene and each column corresponds to a sample. The rownames and colnames of refCor should be present in the rownames of emat. For the p-value calculation, it is important that emat include all measured genes, not just those in refCor.

groupVec
Optional vector indicating the group to which group each sample belongs. If not provided, the function assumes all samples belong to the same group.

refEmat
Optional expression matrix for calculating co-expression for the reference, with the same organization as emat. Only used if refCor is not provided.

nPerm
Number of permutations for assessing statistical significance.

geneNames
Optional vector indicating a subset of genes in refCor, emat, and/or refEmat to use for calculating the CCD.

dopar
Logical indicating whether to process features in parallel. Prior to calling calcCCD(), use registerDoParallel() to register the parallel backend, followed by set.seed() to make the p-values reproducible.

Value

A data frame with columns for group name, CCD, and p-value.

See also

Examples

if (FALSE) {
library('deltaccd')
library('doParallel')
library('doRNG')

registerDoParallel(cores = 2)
set.seed(35813)

refCor = getRefCor()
ccdResult = calcCCD(refCor, GSE19188$emat, GSE19188$groupVec, dopar = TRUE)
deltaCcdResult = calcDeltaCCD(refCor, GSE19188$emat, GSE19188$groupVec,
                              'non-tumor', dopar = TRUE)

pRef = plotRefHeatmap(refCor)
pTest = plotHeatmap(rownames(refCor), GSE19188$emat, GSE19188$groupVec)
}