Retrieve the reference correlation matrix for circadian gene co-expression.

The pan-tissue reference matrix is based on a fixed-effects meta-analysis of eight circadian transcriptome datasets from mice, as described in Shilts et al. 2018. The human blood reference matrix is based an analysis of three microarray datasets (manuscript in preparation).

getRefCor(
  species = c("human", "mouse"),
  tissue = c("pan", "blood"),
  useEntrezGeneId = TRUE
)

Arguments

species
Currently either 'human' or 'mouse'. Only affects the row and column names of the correlation matrix, not the actual values.

tissue
One of either 'pan' or 'blood'.

useEntrezGeneId
If FALSE, row and column names of correlation matrix will correspond to gene symbols (e.g., PER2).

Value

A matrix of Spearman correlation values.

See also

Examples

if (FALSE) {
library('deltaccd')
library('doParallel')
library('doRNG')

registerDoParallel(cores = 2)
set.seed(35813)

refCor = getRefCor()
ccdResult = calcCCD(refCor, GSE19188$emat, GSE19188$groupVec, dopar = TRUE)
deltaCcdResult = calcDeltaCCD(
  refCor, GSE19188$emat, GSE19188$groupVec, 'non-tumor', dopar = TRUE)

pRef = plotRefHeatmap(refCor)
pTest = plotHeatmap(rownames(refCor), GSE19188$emat, GSE19188$groupVec)
}